ConnectionからRequestを分離
